+++ examples/data-sources/nxos_feature_ospfv3/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_feature_ospfv3" "example" {
+}