+++ examples/resources/nxos_feature_ospfv3/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_ospfv3.example "sys/fm/ospfv3"