+++ examples/resources/nxos_feature_ospfv3/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_ospfv3" "example" {
+  admin_state = "enabled"
+}