+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add nxos_feature_lacp resource and data source
 - Add nxos_feature_macsec resource and data source
 - Add nxos_feature_netflow resource and data source
+- Add nxos_feature_ospfv3 resource and data source
 
 ## 0.3.0